Added SVG icons and tidied up code and styling a bit more.
