--- components/home-layout.tsx
@@ -1,14 +1,0 @@
-import Head from 'next/head'
-import React from 'react'
-import styles from './layout.module.css'
-
-export default function HomeLayout({ children }: { children: React.ReactNode }) {
-  return (
-    <div className="font-body text-black text-opacity-80 bg-beige">
-      <Head>
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      {children}
-    </div>
-  )
-}
