+++ components/home/landing-screen.tsx
@@ -1,9 +1,7 @@
-import React, { Fragment } from 'react'
+import React from 'react'
 import { LandingScreenEntry } from '../../lib/content'
 import { FaceCanvas } from './dynamic/face-canvas'
-import LinkedInIcon from '../../content/icons/linkedin.svg'
-import InstagramIcon from '../../content/icons/instagram.svg'
-import GitHubIcon from '../../content/icons/github.svg'
+import { SocialMediaLinks } from './social-media-links'
 
 type LandingScreenProps = {
     entry: LandingScreenEntry
@@ -31,29 +29,5 @@ export const LandingScreen = ({ entry } : LandingScreenProps) => {
                 </section>
             </div>
         </>
-    )
-}
-
-type SocialMediaLinksProps = {
-    linkedIn: string
-    gitHub: string
-    instagram: string
-}
-
-const SocialMediaLinks = ({linkedIn, gitHub, instagram} : SocialMediaLinksProps) => {
-    return (
-        <ul className="list-none inline-grid grid-flow-col auto-cols-max gap-5">
-            {linkedIn && <li>
-                <a href={linkedIn}><LinkedInIcon className="h-7 w-7 lg:h-8 lg:w-8 fill-current"/></a>
-            </li>}
-
-            { instagram && <li>
-                <a href={instagram}><InstagramIcon className="h-7 w-7 lg:h-8 lg:w-8 fill-current"/></a>
-            </li> }
-            
-            { gitHub && <li>
-                <a href={gitHub}><GitHubIcon className="h-7 w-7 lg:h-8 lg:w-8 fill-current"/></a>
-            </li> }
-        </ul>
     )
 }
