+++ content/home/landing-screen.md
@@ -2,6 +2,8 @@
 author: Ferugi El Heri
 title: Ferugi
 tagline: Hi, I’m a full-stack developer living in Brighton, UK.
+description:
+  Ferugi El Heri is a full-stack software developer skilled in .NET, React, Angular, and many more technologies. Passionate about software development, powerlifting, and life in Brighton, UK.
 contactLinks:
   email: me@ferugi.com
   linkedIn: https://www.linkedin.com/in/ferugi
