+++ netlify-cms-config.yml
@@ -13,6 +13,9 @@ collections:
     name: "pages"
     files:
 
+    # TODO: Revaluate the structure, and even use of this
+    # Social Media links for example should be its own collection
+    # As well as site & author details
     - label: "Landing Screen"
       name: "landingScreen"
       file: "content/home/landing-screen.md"
@@ -20,6 +23,7 @@ collections:
         - { label: "Author", name: "author", widget: "string"}
         - { label: "Title", name: "title", widget: "string"}
         - { label: "Tagline", name: "tagline", widget: "string"}
+        - { label: "Site Description", name: "description", widget: "string"}
         - label: "Contact Links"
           name: "contactLinks"
           widget: "object"
@@ -67,6 +71,7 @@ collections:
       - { label: "Facebook Comment Link", name: "fbcommentlink", widget: "hidden", required: false }
       - { label: "Categories", name: "categories", widget: "list", required: false }
       - { label: "Tags", name: "tags", widget: "list", required: false}
+      - { label: "Post Description", name: "description", widget: "string"}
       - { label: "Body", name: "body", widget: "markdown" }
 
   - label: "CV Experiences"
