+++ pages/blog/[id].tsx
@@ -1,4 +1,4 @@
-import ArticleLayout from "../../components/article-layout"
+import Layout from "../../components/layout"
 import Head from 'next/head'
 import Date from '../../components/date'
 import { GetStaticProps, GetStaticPaths } from 'next'
@@ -9,20 +9,23 @@ import FacebookComments from "../../components/fb-comments"
 
 export default function Post({ post }: { post: WithoutDate<BlogPostEntry> }) {
   return (
-    <ArticleLayout>
+    <Layout>
       <Head>
         <title>{post.title}</title>
+        <meta name="description" content={post.description} />
       </Head>
+    <main className="container max-w-text mx-auto py-6 sm:py-10 md:py-14 px-4 sm:px-5 md:px-6">
       <article>
-        <header className="mb-5">
-          <h1 className={styles.postTitle}>{post.title}</h1>
-          <Date className="font-body text-xs" dateString={post.date as string} />
-        </header>
-        <div className={styles.postBody} dangerouslySetInnerHTML={{ __html: post.body }} />
-      </article>
-      { !!post.fbcommentlink &&  
-        <FacebookComments commentUrl={post.fbcommentlink} orderBy="reverse_time" width={635}/> }
-    </ArticleLayout>
+          <header className="mb-5">
+            <h1 className={styles.postTitle}>{post.title}</h1>
+            <Date className="font-body text-xs" dateString={post.date as string} />
+          </header>
+          <div className={styles.postBody} dangerouslySetInnerHTML={{ __html: post.body }} />
+        </article>
+        { !!post.fbcommentlink &&  
+          <FacebookComments commentUrl={post.fbcommentlink} orderBy="reverse_time" width={635}/> }
+    </main>
+    </Layout>
   )
 }
