+++ public/manifest.json
@@ -1,0 +1,11 @@
+{
+    "name": "Starter",
+    "short_name": "Starter",
+    "icons": [{
+        "src": "google-touch-icon-512.png",
+        "sizes": "512x512"
+    }],
+    "background_color": "#ffffff",
+    "theme_color": "#ffffff",
+    "display": "fullscreen"
+}
